@@ -25,6 +25,7 @@
 #else
   jsCodeLocation = [CodePush bundleURL];
 #endif
+//  jsCodeLocation = [CodePush bundleURL];
   
   RCTRootView *rootView = [[RCTRootView alloc] initWithBundleURL:jsCodeLocation
                                                       moduleName:@"YrcnAppProject"
@@ -39,9 +40,9 @@
   [self.window makeKeyAndVisible];
   //如果已经得到授权，就直接添加本地通知，否则申请询问授权
   if ([[UIApplication sharedApplication]currentUserNotificationSettings].types!=UIUserNotificationTypeNone) {
-    NSLog(@"取得通知授权");
+    //NSLog(@"取得通知授权");
   }else{
-    NSLog(@"未取得通知授权");
+    //NSLog(@"未取得通知授权");
     [[UIApplication sharedApplication]registerUserNotificationSettings:[UIUserNotificationSettings settingsForTypes:UIUserNotificationTypeAlert|UIUserNotificationTypeBadge|UIUserNotificationTypeSound  categories:nil]];
   }
   //
