@@ -1,7 +1,8 @@
 //
 var configs = {
     AppName: '今天做了啥',
-    IOS_APP_URL: 'https://itunes.apple.com/us/app/jin-tian-zuo-le-sha/id1167259165?l=zh&ls=1&mt=8'
+    IOS_APP_URL: 'https://itunes.apple.com/us/app/jin-tian-zuo-le-sha/id1167259165?l=zh&ls=1&mt=8',
+    AS_KEY_WORKING_LOG : "AS_工作日志$$",//存储系统 工作日志
 };
 //
 module.exports = configs;
