@@ -18,54 +18,59 @@ var NavigatorToday = require('./../navigator/NavigatorToday.js');
 var NavigatorYesterday = require('./../navigator/NavigatorYesterday.js');
 var NavigatorLlg = require('./../navigator/NavigatorLlg.js');
 var NavigatorSettings = require('./../navigator/NavigatorSettings.js');
+
+
+var ScrollViewToday = require('./../scrollview/ScrollViewToday.js');
+var ScrollViewYesterday = require('./../scrollview/ScrollViewYesterday.js');
+var ScrollViewLlg = require('./../scrollview/ScrollViewLlg.js');
+var SettingsView = require('./../view/ViewSettings.js');
 //
 var TabBarIndex = React.createClass({
     getInitialState: function(){
         var _this = this;
         return {
-            selectedTab: 'todayIcon',
+            selectedTab: this.props.selectedTab||'todayIcon',
         };
     },
     render: function(){
-        //console.log("TabBarIndex.js render");
+        //
+        var _this = this;
+        global.YrcnApp.now.$TabBarIndex = this;
+        global.YrcnApp.components.StatusBar.setBarStyle('dark-content',false);
         return (
             <TabBarIOS barTintColor={'#f6f6f7'} tintColor={'#01bbfc'} translucent={true} style={styles.container}>
                 <TabBarIOS.Item
-                    icon={require('image!hotLife')}
-                    selectedIcon={require('image!hotLife')}
+                    icon={require('../images/hotLife.png')}
+                    selectedIcon={require('../images/hotLife.png')}
                     selected={this.state.selectedTab === 'todayIcon'}
                     onPress={(item)=>{this._onPressTabItem('todayIcon')}}
                     title="今天">
-                    <NavigatorToday />
+                    <ScrollViewToday />
                 </TabBarIOS.Item>
                 <TabBarIOS.Item
-                    icon={require('image!yesterdayIcon')}
-                    selectedIcon={require('image!yesterdayIcon')}
+                    icon={require('../images/yesterdayIcon.png')}
+                    selectedIcon={require('../images/yesterdayIcon.png')}
                     selected={this.state.selectedTab === 'yesterdayIcon'}
                     onPress={(item)=>{this._onPressTabItem('yesterdayIcon')}}
                     title="昨天">
-                    <NavigatorYesterday
-                        NavigatorRoot_route={this.props.NavigatorRoot_route}
-                        NavigatorRoot_navigator={this.props.NavigatorRoot_navigator}/>
+                    <ScrollViewYesterday contentDay = {global.YrcnApp.utils.yesterdayDate()}
+                                    backgroundColor = "#ffffff"/>
                 </TabBarIOS.Item>
                 <TabBarIOS.Item
-                    icon={require('image!llgIcon')}
-                    selectedIcon={require('image!llgIcon')}
+                    icon={require('../images/llgIcon.png')}
+                    selectedIcon={require('../images/llgIcon.png')}
                     selected={this.state.selectedTab === 'llgIcon'}
                     onPress={(item)=>{this._onPressTabItem('llgIcon')}}
                     title="以前">
-                    <NavigatorLlg
-                        NavigatorRoot_route={this.props.NavigatorRoot_route}
-                        NavigatorRoot_navigator={this.props.NavigatorRoot_navigator}/>
+                    <ScrollViewLlg />
                 </TabBarIOS.Item>
                 <TabBarIOS.Item
-                    icon={require('image!myLife')}
+                    icon={require('../images/myLife.png')}
+                    selected={require('../images/myLife.png')}
                     selected={this.state.selectedTab === 'myLife'}
                     onPress={()=>{this.setState({selectedTab: 'myLife'})}}
                     title="设置">
-                    <NavigatorSettings
-                        NavigatorRoot_route={this.props.NavigatorRoot_route}
-                        NavigatorRoot_navigator={this.props.NavigatorRoot_navigator}/>
+                    <SettingsView />
                 </TabBarIOS.Item>
             </TabBarIOS>
         );
