@@ -24,28 +24,28 @@ import {
 var ButtonsBox = require('../../component/ButtonsBox.js');
 var ACViewBox = require('../../component/ACViewBox.js')
 //
-/**
- * 定义属性：
- */
-var GuideIndexView = React.createClass({
-    _vars:{
-    },
-    getDefaultProps: function(){
-        return ({
 
-        });
-    },
-    getInitialState: function(){
-        return ({
-            isPressingWelcome: false,
-            isPressingStart: false,
-            isPressingLogin: false,
-            isLogined: "0",
-        });
-    },
+class GuideIndexView  extends React.Component {
+    constructor(props) {
+        super(props);
+        this._onPressWelcome = this._onPressWelcome.bind(this);
+    }
+    static navigationOptions = {
+        drawerLabel: 'Home',
+    };
+    _vars={
+    }
+    static defaultProps={
+    }
+    state={
+        isPressingWelcome: false,
+        isPressingStart: false,
+        isPressingLogin: false,
+        isLogined: "0",
+    }
     //在初始化渲染执行之后立刻调用一次，仅客户端有效（服务器端不会调用）。在生命周期中的这个时间点，组件拥有一个 DOM 展现，你可以通过 this.getDOMNode() 来获取相应 DOM 节点。
     //如果想和其它 JavaScript 框架集成，使用 setTimeout 或者 setInterval 来设置定时器，或者发送 AJAX 请求，可以在该方法中执行这些操作。
-    componentDidMount: function(){
+    componentDidMount(){
         var _this = this;
         global.YrcnApp.utils.getLoginInfo(function(loginInfo){
             if(loginInfo && loginInfo.userLogin){
@@ -58,18 +58,19 @@ var GuideIndexView = React.createClass({
                 });
             }
         })
-    },
+    }
     //在组件从 DOM 中移除的时候立刻被调用。
     //在该方法中执行任何必要的清理，比如无效的定时器，或者清除在 componentDidMount 中创建的 DOM 元素。
-    componentWillUnmount: function(){
+    componentWillUnmount(){
         var _this = this;
-    },
+    }
     //
-    render: function(){
+    render(){
         var _this = this;
         global.YrcnApp.components.StatusBar.setHidden(false,'slide');
         global.YrcnApp.components.StatusBar.setBarStyle('light-content',false);
-        this.props.parent.hideLeftButton();
+        //global.YrcnApp.root_navigate = _this.props.navigation.navigate;
+        //this.props.parent.hideLeftButton();
         //
         return (
             <View style={[styles.container]}>
@@ -82,7 +83,9 @@ var GuideIndexView = React.createClass({
                             if(_this.state.isLogined == "1"){
                                 return (
                                     <ButtonsBox marginBottom={0}>
-                                        <ButtonsBox.Button btnText={"欢迎回来，"+global.YrcnApp.loginUser.userInfo.niCheng} onPress={_this._onPressWelcome} isPressing={_this.state.isPressingWelcome}/>
+                                        <ButtonsBox.Button btnText={"欢迎回来，"+global.YrcnApp.loginUser.userInfo.niCheng}
+                                                           onPress={_this._onPressWelcome}
+                                                           isPressing={_this.state.isPressingWelcome}/>
                                     </ButtonsBox>
                                 );
                             }else if(_this.state.isLogined == "2"){
@@ -102,18 +105,17 @@ var GuideIndexView = React.createClass({
                 </View>
             </View>
         );
-    },
-    _onPressWelcome: function(){
-        global.YrcnApp.now.rootNavigator.replace({name:'TabBarIndex'});
-    },
-    _onPressStart: function () {
-        //this.props.parent_navigator.push({name:'GuideProfessionView',title:'Today'});
-        this.props.parent_navigator.push({name:'RegisterIndexView',title:'Today'});
-    },
-    _onPressLogin: function () {
-        this.props.parent_navigator.push({name:'LoginIndexView',title:"Today"});
-    },
-});
+    }
+    _onPressWelcome(){
+        YrcnApp.now.$ViewRoot.setState({viewName:'TabBarIndex'});
+    }
+    _onPressStart(){
+        YrcnApp.now.$ViewRoot.setState({viewName:'RegisterIndexView'});
+    }
+    _onPressLogin(){
+        YrcnApp.now.$ViewRoot.setState({viewName:'LoginIndexView'});
+    }
+}
 //
 module.exports = GuideIndexView;
 //
